Updated Login component. Moved Auth inside container. Move session handling to helper.
